pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nicegui-sso-demo"
version = "0.1.0"
description = "Demonstration of NiceGUI application with SSO integration"
authors = [
    {name = "Raül Cruz", email = "rcruz@pintaluba.com"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nicegui>=1.4.0",
    "pyjwt>=2.8.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "taskipy>=1.12.0",
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.taskipy.tasks]
dev = "nicegui-pack --recursive main.py"
start = "python main.py"
docker-build = "docker compose build"
test = "pytest"
lint = "ruff check . && black --check ."
format = "black . && ruff check . --fix"
